+++ tests/phpunit/tests/add-admin-css.php
@@ -188,11 +188,11 @@ class Add_Admin_CSS_Test extends WP_UnitTestCase {
 	}
 
 	public function test_plugin_framework_class_name() {
-		$this->assertTrue( class_exists( 'c2c_Plugin_067' ) );
+		$this->assertTrue( class_exists( 'c2c_Plugin_068' ) );
 	}
 
 	public function test_plugin_framework_version() {
-		$this->assertEquals( '067', $this->obj->c2c_plugin_version() );
+		$this->assertEquals( '068', $this->obj->c2c_plugin_version() );
 	}
 
 	public function test_version() {